requirements.txt
pandas>=1.0.0
numpy>=1.18.0
scikit-learn>=0.22.0
matplotlib>=3.1.0
seaborn>=0.10.0
scipy>=1.4.0
joblib>=0.14.0
openai>=1.0.0
groq>=0.4.0
python-dotenv>=1.0.0
